refactor: View와 API Controller 분리
