+++ src/main/java/roomescape/controller/RoomescapeController.java
@@ -1,17 +1,21 @@
-package roomescape;
+package roomescape.controller;
 
 import jakarta.validation.Valid;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.Reservation;
+import roomescape.ReservationManager;
+import roomescape.ReservationRequest;
+import roomescape.Reservations;
 
-@Controller
+@RestController
 public class RoomescapeController {
 
     private final Reservations reservations;
@@ -20,16 +24,6 @@ public class RoomescapeController {
     public RoomescapeController(final Reservations reservations, final ReservationManager reservationManager) {
         this.reservations = reservations;
         this.reservationManager = reservationManager;
-    }
-
-    @GetMapping("/admin")
-    public String getAdminPage() {
-        return "admin/index";
-    }
-
-    @GetMapping("/admin/reservation")
-    public String getReservationPage() {
-        return "admin/reservation-legacy";
     }
 
     @GetMapping("/reservations")
